Fix ScrollableView ContentSize Calculation
The previous calculation just ignored one part of the size from `intrinsicContentSize`, but in our case the result was larger than the scrollView's width. Just decreasing the width makes it fit, but this could also make the result bigger in the other axis (think text: If you decrease the width of the text box, the text will just expand vertically).

This was fixed by using sizeThatFits instead of intrinsicContentSize and passing the correct bounding parameter (width or height, depending on scrollview axis).